Refactor HUB root URLs for speed
