+++ ultralytics/hub/__init__.py
@@ -5,12 +5,11 @@ import requests
 from ultralytics.data.utils import HUBDatasetStats
 from ultralytics.hub.auth import Auth
 from ultralytics.hub.session import HUBTrainingSession
-from ultralytics.hub.utils import HUB_API_ROOT, HUB_WEB_ROOT, PREFIX, events
-from ultralytics.utils import LOGGER, SETTINGS, checks
+from ultralytics.hub.utils import PREFIX, events
+from ultralytics.utils import LOGGER, SETTINGS, checks, HUB_API_ROOT, HUB_WEB_ROOT
 
 __all__ = (
     "PREFIX",
-    "HUB_WEB_ROOT",
     "HUBTrainingSession",
     "login",
     "logout",
