+++ ultralytics/hub/auth.py
@@ -2,8 +2,8 @@
 
 import requests
 
-from ultralytics.hub.utils import HUB_API_ROOT, HUB_WEB_ROOT, PREFIX, request_with_credentials
-from ultralytics.utils import IS_COLAB, LOGGER, SETTINGS, emojis
+from ultralytics.hub.utils import PREFIX, request_with_credentials
+from ultralytics.utils import IS_COLAB, LOGGER, SETTINGS, emojis, HUB_API_ROOT, HUB_WEB_ROOT
 
 API_KEY_URL = f"{HUB_WEB_ROOT}/settings?tab=api+keys"
