+++ ultralytics/utils/callbacks/hub.py
@@ -3,8 +3,8 @@
 import json
 from time import time
 
-from ultralytics.hub import HUB_WEB_ROOT, PREFIX, HUBTrainingSession, events
-from ultralytics.utils import LOGGER, RANK, SETTINGS
+from ultralytics.hub import PREFIX, HUBTrainingSession, events
+from ultralytics.utils import LOGGER, RANK, SETTINGS, HUB_WEB_ROOT
 
 
 def on_pretrain_routine_start(trainer):
